refactor: keep Legacy/CRD/Hybrid providers in single files
